change how RUN_SERIAL is implemented
The original implementation of the RUN_SERIAL test property
worked by having such a test consume all available processors.
This strategy causes unnecessary timeouts and delays when used
in conjunction with the new --max-load feature.

So instead of artificially inflating the number of processors
that a RUN_SERIAL test requires, we now use a boolean ivar to ensure
that such a test is not run in parallel with any other tests.