QtAutogen: Use PRE_BUILD in Visual Studio generators even with rcc
In commit v3.2.0-rc1~480^2 (QtAutogen: Regenerate qrc files if their
input changes, 2014-09-17) we stopped using the PRE_BUILD step in
Visual Studio generators when rcc is on in order to run it as a
custom command with dependencies.  Our parent commit switched back
to always running the command, so now it can be in PRE_BUILD again.